Added confirmation to delete page

+++ delete.php
@@ -2,6 +2,12 @@
     require("Data/PHP/DB/DBmanager.php");
 
     $conn = createConn("gamemanager", "yQYlpIQ9tyEVZeFV", "spellen");
-    DBcommand($conn, "DELETE FROM planning WHERE `planning` . `id` = :id", [":id" => $_POST['id']]);
 
+    if($_POST['confirmation'] == 'YES I WANT TO DELETE THIS') {
+        DBcommand($conn, "DELETE FROM planning WHERE `planning` . `id` = :id", [":id" => $_POST['id']]);
+    
+        header("location: index.php");
+    } else {
+        header("location: index.php");
+    }
 ?>
